refactoring the code.
