+++ python/tk_multi_workfiles/file_form_base.py
@@ -36,6 +36,7 @@ from .scene_operation import get_current_path, SAVE_FILE_AS_ACTION
 from .file_item import FileItem
 from .work_area import WorkArea
 from .actions.new_task_action import NewTaskAction
+from .actions.file_action import FileAction
 from .user_cache import g_user_cache
 from .util import monitor_qobject_lifetime, resolve_filters, get_sg_entity_name_field
 from .step_list_filter import get_saved_step_filter
@@ -56,6 +57,8 @@ class FileFormBase(QtGui.QWidget):
         QtGui.QWidget.__init__(self, parent)
 
         self._current_file = None
+        self._navigating = False
+        self._exit_code = QtGui.QDialog.Rejected
 
         # create a single instance of the task manager that manages all
         # asynchrounous work/tasks.
@@ -83,6 +86,27 @@ class FileFormBase(QtGui.QWidget):
             osx_f5_refresh_action.triggered.connect(self._on_refresh_triggered)
             self.addAction(osx_f5_refresh_action)
 
+        app = sgtk.platform.current_bundle()
+
+    def _do_init(self):
+        """
+        """
+
+        # set up the UI
+        self._ui = self.init_ui_file()
+        self._ui.setupUi(self)
+
+        # hook up signals on controls:
+        self._ui.cancel_btn.clicked.connect(self._on_cancel)
+        self._ui.browser.create_new_task.connect(self._on_create_new_task)
+        self._ui.browser.work_area_changed.connect(self._on_browser_work_area_changed)
+        self._ui.browser.step_filter_changed.connect(self._apply_step_filtering)
+        self._ui.nav.navigate.connect(self._on_navigate)
+        self._ui.nav.home_clicked.connect(self._on_navigate_home)
+
+    def init_ui_file(self):
+        raise NotImplementedError
+
     def closeEvent(self, event):
         """
         Overriden method triggered when the widget is closed.  Cleans up as much as possible
@@ -90,6 +114,9 @@ class FileFormBase(QtGui.QWidget):
 
         :param event:   Close event
         """
+
+        # clean up the browser:
+        self._ui.browser.shut_down()
 
         # clear up the various data models:
         if self._file_model:
@@ -284,6 +311,62 @@ class FileFormBase(QtGui.QWidget):
         app.log_debug("Path cache up to date!")
         self._refresh_all_async()
 
+    def _on_browser_work_area_changed(self, entity, breadcrumbs):
+        """
+        Slot triggered whenever the work area is changed in the browser.
+        """
+        env_details = None
+        if entity:
+            # (AD) - we need to build a context and construct the environment details
+            # instance for it but this may be slow enough that we should cache it!
+            # Keep an eye on it and consider threading if it's noticeably slow!
+            app = sgtk.platform.current_bundle()
+            context = app.sgtk.context_from_entity_dictionary(entity)
+            try:
+                env_details = WorkArea(context)
+            except sgtk.TankError:
+                # We can ignore the error reporting here. The browser is already
+                # updating it's various file views and they will display the same
+                # error. Which is good, because file open dialog doesn't have a
+                # widget dedicated to error reporting.
+                env_details = None
+
+        if not self._navigating:
+            destination_label = breadcrumbs[-1].label if breadcrumbs else "..."
+            self._ui.nav.add_destination(destination_label, breadcrumbs)
+        self._ui.breadcrumbs.set(breadcrumbs)
+
+        return env_details
+
+    def _on_navigate(self, breadcrumb_trail):
+        """
+        """
+        if not breadcrumb_trail:
+            return
+
+        # awesome, just navigate to the breadcrumbs:
+        self._ui.breadcrumbs.set(breadcrumb_trail)
+        self._navigating = True
+        try:
+            self._ui.browser.navigate_to(breadcrumb_trail)
+        finally:
+            self._navigating = False
+
+    def _on_navigate_home(self):
+        """
+        Navigate to the current work area
+        """
+        # navigate to the current work area in the browser:
+        app = sgtk.platform.current_bundle()
+        self._ui.browser.select_work_area(app.context)
+
+    def _on_cancel(self):
+        """
+        Called when the cancel button is clicked
+        """
+        self._exit_code = QtGui.QDialog.Rejected
+        self.close()
+
     def _refresh_all_async(self):
         """
         Asynchrounously refresh all models.
@@ -388,3 +471,35 @@ class FileFormBase(QtGui.QWidget):
         for _, _, model in self._entity_models:
             if model.supports_step_filtering:
                 model.update_filters(step_filter)
+
+    def _perform_action(self, action):
+        """
+        """
+        if not action:
+            return
+
+        # some debug:
+        app = sgtk.platform.current_bundle()
+        if isinstance(action, FileAction) and action.file:
+            app.log_debug(
+                "Performing action '%s' on file '%s, v%03d'"
+                % (action.label, action.file.name, action.file.version)
+            )
+        else:
+            app.log_debug("Performing action '%s'" % action.label)
+
+        # execute the action:
+        close_dialog = action.execute(self)
+
+        # if this is successful then close the form:
+        if close_dialog:
+            self._exit_code = QtGui.QDialog.Accepted
+            self.close()
+        else:
+            # refresh all models in case something changed as a result of
+            # the action (especially important with custom actions):
+            self._refresh_all_async()
+
+    @property
+    def exit_code(self):
+        return self._exit_code
